lean el comentario este comentario putos
el programa ya colisiona contra las paredes del mundo, contra los enemigos, de alguna forma ataca y mata a los enemigos.
ya es capaz de moverse con las flechas.
aun no logro que la cantidad de vidas del jugador cambie (Graficamente) y tampoco se como corregir el glitch que se da al chocar contra un enemigo puede que sea porque son las 2 de la mañana

+++ src/model/Enemy.java
@@ -74,7 +74,7 @@ public class Enemy extends Character{
 		return lives;
 	}
 
-	public void setLives(short lives) {
+	public void setLives(int lives) {
 		this.lives = lives;
 	}
